refactor: split palette integration monoliths
