+++ crates/axon-code-index/src/indexer.rs
@@ -73,81 +73,15 @@ async fn reindex_changed_files_inner(
         summary.removed_files += 1;
     }
 
-    let batch_size = changed_file_batch_size();
-    let total_batches = manifest.files.len().div_ceil(batch_size);
-    for (batch_index, batch) in manifest.files.chunks(batch_size).enumerate() {
-        let batch_number = batch_index + 1;
-        let mut prepared = Vec::new();
-        let first_path = batch
-            .first()
-            .map(|entry| entry.relative_path.as_str())
-            .unwrap_or_default();
-        let last_path = batch
-            .last()
-            .map(|entry| entry.relative_path.as_str())
-            .unwrap_or_default();
-        tracing::info!(
-            project_key = %identity.project_key,
-            generation,
-            batch = batch_number,
-            total_batches,
-            batch_files = batch.len(),
-            first_path,
-            last_path,
-            "code-search reindex batch start"
-        );
-
-        for entry in batch {
-            store
-                .mark_file_pending(identity, &entry.relative_path)
-                .await?;
-            cleanup_paths.push(entry.relative_path.clone());
-            if entry.size_bytes == 0 {
-                continue;
-            }
-
-            let Some(doc) = prepare_local_code_doc(identity, entry, generation).await? else {
-                continue;
-            };
-            prepared.push(doc);
-        }
-
-        if let Some(cfg) = cfg
-            && !prepared.is_empty()
-        {
-            tracing::info!(
-                project_key = %identity.project_key,
-                generation,
-                batch = batch_number,
-                total_batches,
-                prepared_docs = prepared.len(),
-                "code-search reindex batch embed start"
-            );
-            embed_prepared_docs(cfg, prepared, None)
-                .await
-                .map_err(|err| anyhow::anyhow!("code search embed failed: {err}"))?
-                .require_success("code search embed")
-                .map_err(|err| anyhow::anyhow!(err))?;
-            tracing::info!(
-                project_key = %identity.project_key,
-                generation,
-                batch = batch_number,
-                total_batches,
-                "code-search reindex batch embed done"
-            );
-        }
-
-        for entry in batch {
-            store.mark_file_indexed(identity, entry, generation).await?;
-        }
-        tracing::info!(
-            project_key = %identity.project_key,
-            generation,
-            batch = batch_number,
-            total_batches,
-            "code-search reindex batch marked indexed"
-        );
-    }
+    reindex_manifest_batches(
+        cfg,
+        store,
+        identity,
+        manifest,
+        generation,
+        &mut cleanup_paths,
+    )
+    .await?;
     summary.indexed_files = diff.added.len() + diff.modified.len();
 
     cleanup_paths.sort();
@@ -180,6 +114,151 @@ async fn reindex_changed_files_inner(
         "code-search reindex cleanup done"
     );
     Ok(summary)
+}
+
+async fn reindex_manifest_batches(
+    cfg: Option<&Config>,
+    store: &CodeIndexStore,
+    identity: &CodeIndexIdentity,
+    manifest: &ManifestSnapshot,
+    generation: i64,
+    cleanup_paths: &mut Vec<String>,
+) -> anyhow::Result<()> {
+    let batch_size = changed_file_batch_size();
+    let total_batches = manifest.files.len().div_ceil(batch_size);
+    for (batch_index, batch) in manifest.files.chunks(batch_size).enumerate() {
+        let batch_number = batch_index + 1;
+        log_reindex_batch_start(identity, generation, batch_number, total_batches, batch);
+        let prepared =
+            prepare_reindex_batch(store, identity, generation, cleanup_paths, batch).await?;
+        embed_reindex_batch(
+            cfg,
+            identity,
+            generation,
+            batch_number,
+            total_batches,
+            prepared,
+        )
+        .await?;
+        mark_reindex_batch_indexed(
+            store,
+            identity,
+            generation,
+            batch_number,
+            total_batches,
+            batch,
+        )
+        .await?;
+    }
+    Ok(())
+}
+
+fn log_reindex_batch_start(
+    identity: &CodeIndexIdentity,
+    generation: i64,
+    batch_number: usize,
+    total_batches: usize,
+    batch: &[FileManifestEntry],
+) {
+    let first_path = batch
+        .first()
+        .map(|entry| entry.relative_path.as_str())
+        .unwrap_or_default();
+    let last_path = batch
+        .last()
+        .map(|entry| entry.relative_path.as_str())
+        .unwrap_or_default();
+    tracing::info!(
+        project_key = %identity.project_key,
+        generation,
+        batch = batch_number,
+        total_batches,
+        batch_files = batch.len(),
+        first_path,
+        last_path,
+        "code-search reindex batch start"
+    );
+}
+
+async fn prepare_reindex_batch(
+    store: &CodeIndexStore,
+    identity: &CodeIndexIdentity,
+    generation: i64,
+    cleanup_paths: &mut Vec<String>,
+    batch: &[FileManifestEntry],
+) -> anyhow::Result<Vec<axon_vector::ops::PreparedDoc>> {
+    let mut prepared = Vec::new();
+    for entry in batch {
+        store
+            .mark_file_pending(identity, &entry.relative_path)
+            .await?;
+        cleanup_paths.push(entry.relative_path.clone());
+        if entry.size_bytes == 0 {
+            continue;
+        }
+        if let Some(doc) = prepare_local_code_doc(identity, entry, generation).await? {
+            prepared.push(doc);
+        }
+    }
+    Ok(prepared)
+}
+
+async fn embed_reindex_batch(
+    cfg: Option<&Config>,
+    identity: &CodeIndexIdentity,
+    generation: i64,
+    batch_number: usize,
+    total_batches: usize,
+    prepared: Vec<axon_vector::ops::PreparedDoc>,
+) -> anyhow::Result<()> {
+    let Some(cfg) = cfg else {
+        return Ok(());
+    };
+    if prepared.is_empty() {
+        return Ok(());
+    }
+    tracing::info!(
+        project_key = %identity.project_key,
+        generation,
+        batch = batch_number,
+        total_batches,
+        prepared_docs = prepared.len(),
+        "code-search reindex batch embed start"
+    );
+    embed_prepared_docs(cfg, prepared, None)
+        .await
+        .map_err(|err| anyhow::anyhow!("code search embed failed: {err}"))?
+        .require_success("code search embed")
+        .map_err(|err| anyhow::anyhow!(err))?;
+    tracing::info!(
+        project_key = %identity.project_key,
+        generation,
+        batch = batch_number,
+        total_batches,
+        "code-search reindex batch embed done"
+    );
+    Ok(())
+}
+
+async fn mark_reindex_batch_indexed(
+    store: &CodeIndexStore,
+    identity: &CodeIndexIdentity,
+    generation: i64,
+    batch_number: usize,
+    total_batches: usize,
+    batch: &[FileManifestEntry],
+) -> anyhow::Result<()> {
+    for entry in batch {
+        store.mark_file_indexed(identity, entry, generation).await?;
+    }
+    tracing::info!(
+        project_key = %identity.project_key,
+        generation,
+        batch = batch_number,
+        total_batches,
+        "code-search reindex batch marked indexed"
+    );
+    Ok(())
 }
 
 async fn prepare_local_code_doc(
